+++ src-docs/src/views/app_view.js
@@ -64,7 +64,7 @@ export class AppView extends Component {
         <EuiPageBody>
           <EuiErrorBoundary>
             <GuidePageChrome
-              currentRouteName={currentRoute.name}
+              currentRoute={currentRoute}
               onToggleTheme={toggleTheme}
               selectedTheme={theme}
               onToggleLocale={toggleLocale}